Use xdebug links for messages, move to trait
